Fix : 타입 수정

+++ back/src/main/java/com/back/domain/news/news/service/NewsService.java
@@ -27,7 +27,7 @@ public class NewsService {
         return newsRepository.findAll(pageable);
     }
 
-    public News getNewsById(Integer id) {
+    public News getNewsById(Long id) {
         return newsRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("뉴스를 찾을 수 없습니다."));
     }
